pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wafpass-server"
version = "0.4.0"
description = "WAF++ PASS – API server for persisting and querying scan results"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Artem Lajko" },
    { name = "Sascha Lewandowski" },
    { name = "Sebastian Meyer" },
    { name = "Tim Urlaub" },
]
keywords = [
    "waf++", "iac", "compliance", "security", "cloud", "api",
    "devsecops", "governance", "fastapi", "multi-cloud",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "wafpass-core>=0.4.0",
    "bcrypt>=4.0",
    "PyJWT>=2.8",
    "python-multipart>=0.0.9",
]

[project.urls]
Homepage = "https://waf2p.dev"
Documentation = "https://waf2p.dev/docs/"
Repository = "https://github.com/WAF2p/wafpass-server"
"Bug Tracker" = "https://github.com/WAF2p/wafpass-server/issues"
Community = "https://github.com/WAF2p"

[project.scripts]
wafpass-server = "wafpass_server.main:start"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
    "pytest-cov>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["wafpass_server"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
asyncio_mode = "auto"
